requirements.txt
Scrapy~=2.5.1
oauth2client~=3.0.0
google-api-python-client~=1.5.3
python-dotenv==0.19.2
imgurpython==1.1.7